+++ src/schemas/wechaty-events.ts
