+++ src/schemas/wechaty-options.ts
